+++ packages/step/src/index.ts
@@ -1,4 +1,14 @@
-export type { StepRunner, Step } from './step'
-export { createContextStep, makeStep } from './step'
-export { createTestingStep, makeVerifyStep } from './adapter'
-export type { GenericVerification, VerifyContextOf } from './adapter'
+// Core Execution Engine
+export { createStep } from './execution'
+export type { StepRunner, Step } from './execution'
+
+// Context & State Management
+export { createContextStep } from './context'
+export type { StepBuilder } from './context'
+
+// Verification & Assertions
+export { createVerifyStep } from './verification'
+export type { GenericVerification, VerifyContextOf } from './verification'
+
+// Orchestration & Test Runner Integration
+export { createTestingStep } from './orchestration'
