+++ packages/step/tests/unit/execution.test.ts
@@ -1,32 +1,32 @@
 import { describe, it, expect, vi } from 'vitest'
-import { makeStep } from '@secorto/step'
+import { createStep } from '@secorto/step'
 import type { StepRunner } from '@secorto/step'
 
 const mockRunner: StepRunner = (_title, action) =>
   Promise.resolve(action())
 
-describe('makeStep', () => {
+describe('createStep', () => {
   it('exposes the title on the definition', () => {
-    const step = makeStep(mockRunner, 'MyStep')
+    const step = createStep(mockRunner, 'MyStep')
     const result = step('my step', () => 42)
     expect(result.title).toBe('my step')
   })
 
   it('exposes the action on the definition', () => {
-    const step = makeStep(mockRunner, 'MyStep')
+    const step = createStep(mockRunner, 'MyStep')
     const action = () => 42
     const result = step('my step', action)
     expect(result.action).toBe(action)
   })
 
   it('resolves with the return value of the action', async () => {
-    const step = makeStep(mockRunner, 'MyStep')
+    const step = createStep(mockRunner, 'MyStep')
     const result = await step('my step', () => 42)
     expect(result).toBe(42)
   })
 
   it('resolves with the return value of an async action', async () => {
-    const step = makeStep(mockRunner, 'MyStep')
+    const step = createStep(mockRunner, 'MyStep')
     const result = await step('my step', async () => 'hello')
     expect(result).toBe('hello')
   })
@@ -45,27 +45,27 @@ describe('makeStep', () => {
       return Promise.resolve(action())
     }
 
-    const step = makeStep(runner, 'MyStep')
+    const step = createStep(runner, 'MyStep')
     await step('click the button', () => undefined)
 
     expect(seenTitle).toBe('click the button')
   })
 
   it('is thenable (.then)', async () => {
-    const step = makeStep(mockRunner, 'MyStep')
+    const step = createStep(mockRunner, 'MyStep')
     const result = await step('my step', () => 99).then((v) => v * 2)
     expect(result).toBe(198)
   })
 
   it('is catchable (.catch)', async () => {
     const failingRunner: StepRunner = () => Promise.reject(new Error('fail'))
-    const step = makeStep(failingRunner, 'MyStep')
+    const step = createStep(failingRunner, 'MyStep')
     const result = await step('my step', () => undefined).catch(() => 'caught')
     expect(result).toBe('caught')
   })
 
   it('supports .finally', async () => {
-    const step = makeStep(mockRunner, 'MyStep')
+    const step = createStep(mockRunner, 'MyStep')
     const spy = vi.fn()
     await step('my step', () => 1).finally(spy)
     expect(spy).toHaveBeenCalledOnce()
